MLbackend/requirements.txt
annotated-types==0.7.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contourpy==1.3.0
convokit==3.0.0
cycler==0.12.1
cymem==2.0.8
dill==0.3.9
dnspython==2.7.0
emoji==1.7.0
fonttools==4.54.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
msgpack==1.1.0
msgpack-numpy==0.4.8
murmurhash==1.0.10
networkx==3.3
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
progress==1.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
sentistrength==0.0.9
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
strsimpy==0.2.1
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
